% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIReator.R
\name{FAIReator}
\alias{FAIReator}
\title{FAIR eDNA Template Generator (FAIRe-ator)
This function is to generate FAIR eDNA data templates, based on the sample
types, assay type, and requirement levels of your choice.}
\usage{
FAIReator(
  req_lev = c("M", "HR", "R", "O"),
  sample_type,
  assay_type,
  project_id,
  assay_name,
  projectMetadata_user = NULL,
  sampleMetadata_user = NULL,
  return_wb_object = FALSE
)
}
\arguments{
\item{req_lev}{Requirement level(s) of each fields to include in the data
template. Select one or more from "M", "HR", "R", and "O" (Mandatory,
Highly recommended, Recommended, and Optional, respectively). Default is
c("M", "HR", "R", "O")}

\item{sample_type}{A (list of) Sample type(s). Select one or more from
"Water", "Soil", "Sediment", "Air", "HostAssociated",
"MicrobialMatBiofilm", and "SymbiontAssociated", or "other". "other" will
include all the sample-type-specific fields in the output.}

\item{assay_type}{An approach applied to detect taxon/taxa. Select one
"targeted" (e.g., q/d PCR based detection) or "metabarcoding" (e.g.,
metabarcoding)}

\item{project_id}{A brief, concise project identifier with no spaces or
special characters. This ID will be used in data file names as
'project_id'.}

\item{assay_name}{A brief, concise assay name(s) with no spaces or special
characters. This will be used in data file names as 'assay_name'.}

\item{projectMetadata_user}{(optional) A user-defined field or list of fields
that are not listed in the FAIR eDNA metadata checklist. These fields will
be appended to the end of the projectMetadata.}

\item{sampleMetadata_user}{(optional) A user-defined field or list of fields
that are not listed in the FAIR eDNA metadata checklist. These fields will
be appended to the end of the sampleMetadata.}

\item{return_wb_object}{(optional) return a wb xlsx object for debugging.}
}
\value{
The expected outputs are:
A new folder called "template_<project_id>" (e.g., template_gbr2022) in your
working directory. An Excel file called "<project_id>.xlsx" (e.g.,
gbr2022.xlsx) in the new template folder, containing multiple worksheets in
the file, including README, projectMetadata, sampleMetadata and other
assay_type specific tables (e.g., OTU table if assay_type = "metabarcoding")
Optional returns a 'wb' xlsx object for testing.
}
\description{
FAIR eDNA Template Generator (FAIRe-ator)
This function is to generate FAIR eDNA data templates, based on the sample
types, assay type, and requirement levels of your choice.
}
\examples{
FAIReator(
  sample_type = c("Water", "Sediment"),
  assay_type = "metabarcoding",
  project_id = "gbr2022",
  assay_name = c("MiFish", "crust16S")
)
}
